Can now repesent meetingslonger than 1 hour.

+++ ipad/Classes/MeetingTimerView.h
@@ -15,7 +15,7 @@
 	NSDate *newDate;
 	UIColor *currentTimerColor;
 	NSMutableArray *selectedTimes;
-	float elapsedSeconds;
+	int elapsedSeconds;
 	NSDate *testDate;
 	float hourCounter;
 	float timeToCompare;
